+++ public/admin/_app/immutable/nodes/24.DmWBanc0.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/U5IHDS6h.js";export{e as component};